+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ##### Fixed
 - Failing to clone repeats that only contain calculations.
 - Select one calculations not updating correctly.
+- Irrelevant calculations inside non-first repeat instances run upon load.
 
 [5.2.2] - 2019-04-01
 --------------------